reworked projects under src/ directory
- src/Rhino.Testing: main library source
- src/Rhino.Testing.Tests: mail library tests (includes mixed fixture + attribute init is tested)
- src/Rhino.Testing.Tests.SetupAttribute: tests rhinocore init from attribute
- src/Rhino.Testing.Tests.SetupFixture: tests rhinocore init from fixture
- src/Rhino.Testing.Tests.Utils: utilities used by test assemblies
