--- Rhino.Testing.Inheritence.Tests/SetupFixture.cs
@@ -1,19 +1,0 @@
-using NUnit.Framework;
-
-[SetUpFixture]
-public sealed class SetupFixture : Rhino.Testing.Fixtures.RhinoSetupFixture
-{
-    public override void OneTimeSetup()
-    {
-        base.OneTimeSetup();
-
-        // your custom setup
-    }
-
-    public override void OneTimeTearDown()
-    {
-        base.OneTimeTearDown();
-
-        // you custom teardown
-    }
-}
