+++ src/Rhino.Testing.Tests.SetupAttribute/Rhino.Testing.Configs.xml
