+++ src/Rhino.Testing.Tests.SetupAttribute/Rhino.Testing.Tests.SetupAttribute.csproj
@@ -11,7 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Rhino.Testing.csproj" />
+    <ProjectReference Include="..\Rhino.Testing\Rhino.Testing.csproj" />
   </ItemGroup>
 
   <ItemGroup>
