+++ src/Rhino.Testing.Tests.SetupAttribute/SimpleTest.cs
@@ -3,11 +3,9 @@ using Rhino.Testing.Fixtures;
 
 namespace Rhino.Testing.Attribute.Tests
 {
-
     [RhinoTestFixture]
     public class SimpleTest
     {
-
         [Test]
         public void Test()
         {
@@ -16,6 +14,5 @@ namespace Rhino.Testing.Attribute.Tests
             Assert.That(point.Y, Is.GreaterThan(0));
             Assert.That(point.Z, Is.GreaterThan(0));
         }
-
     }
 }
