+++ src/Rhino.Testing.Tests.SetupFixture/Rhino.Testing.Configs.xml
