+++ src/Rhino.Testing.Tests.SetupFixture/Rhino.Testing.Tests.SetupFixture.csproj
@@ -14,7 +14,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\src\Rhino.Testing.csproj" />
+    <ProjectReference Include="..\Rhino.Testing\Rhino.Testing.csproj" />
+    <ProjectReference Include="..\Rhino.Testing.Tests.Utils\Rhino.Testing.Tests.Utils.csproj" />
   </ItemGroup>
 
   <ItemGroup>
