+++ src/Rhino.Testing.Tests/Rhino.Testing.Configs.xml
