+++ src/Rhino.Testing.Tests/Rhino.Testing.Tests.csproj
@@ -9,12 +9,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
-		<PackageReference Include="NUnit.Analyzers" Version="3.3.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
+    <PackageReference Include="NUnit.Analyzers" Version="3.3.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\src\Rhino.Testing.csproj" />
+    <ProjectReference Include="..\Rhino.Testing\Rhino.Testing.csproj" />
   </ItemGroup>
 
   <ItemGroup>
