+++ src/Rhino.Testing/Configs.cs
