+++ src/Rhino.Testing/Rhino.Testing.csproj
@@ -8,7 +8,7 @@
     <AssemblyName>Rhino.Testing</AssemblyName>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Product>Rhino.Testing</Product>
-    <Version>8.0.17</Version>
+    <Version>8.0.23</Version>
     <PackageVersion>$(Version)-beta</PackageVersion>
     <Title>Unit Testing for Rhino3D</Title>
     <Description>NUnit dotnet testing library for Rhino3D&gt;8</Description>
@@ -17,8 +17,19 @@
     <PackageReadmeFile>README.md</PackageReadmeFile>
   </PropertyGroup>
 
+  <PropertyGroup>
+    <Authors>Robert McNeel and Associates</Authors>
+    <Company>McNeel</Company>
+    <Copyright>Copyright (c) 1997-2024 Robert McNeel and Associates</Copyright>
+    <RepositoryUrl>https://github.com/mcneel/Rhino.Testing.git</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+    <PackageTags>Unit Testing NUnit Rhino Rhino3D</PackageTags>
+    <PackageProjectUrl>https://github.com/mcneel/Rhino.Testing</PackageProjectUrl>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+  </PropertyGroup>
+
   <ItemGroup>
-    <None Include="$(SourceDir)..\README.md" Pack="true" PackagePath="" />
-    <None Include="$(SourceDir)..\art\icon.png" Pack="true" PackagePath="" />
+    <None Include="$(SourceDir)..\..\README.md" Pack="true" PackagePath="" />
+    <None Include="$(SourceDir)..\..\art\icon.png" Pack="true" PackagePath="" />
   </ItemGroup>
 </Project>
