+++ src/Rhino.Testing/_Loaders/PluginLoader.cs
